Add support for Python 3 f-string syntax
Add syntax highlighting for Python 3's f-strings, using the <LexerType>
section's <WordStyle ...> tag attributes: "F STRING", "F CHARACTER", "F
TRIPLE", and "F TRIPLEDOUBLE".

The color scheme for these tags is set to be identical to the non
f-string version of the "STRING" tag.

The "F TRIPLE" and "F TRIPLEDOUBLE" tags are set to follow the "STRING"
tag theming instead of the "TRIPLE"/"TRIPLEDOUBLE" tags, because they
are typically used for string assignments and formatting, whereas the
"TRIPLE" and "TRIPLEDOUBLE" tags are themed to be used for Python
docstring literals and comments.